+++ CoffeShop/CoffeeShop.DataAccess/Repositories/CustomRepositories/CoffeeRepositories/CoffeeRepository.cs
@@ -7,7 +7,7 @@ public class CoffeeRepository : Repository<Coffee>, ICoffeeRepository
 {
     public CoffeeRepository(CoffeeShopContext context) : base(context) { }
 
-    public async Task<Coffee> GetAsync(int id)
+    public async Task<Coffee> GetAsync(long id)
         => await _context.Coffees.FirstOrDefaultAsync(x => x.Id.Equals(id)) 
            ?? throw new NullReferenceException();
 }
